indoor_air_quality_publisher/requirements.txt
setproctitle==1.1.9
hiredis==0.2.0
redis==2.10.5
pyserial==3.0.1
